Rename: lib안에 있는 service 폴더 밖으로 이동
