+++ app/artists/page.tsx
@@ -12,7 +12,7 @@ import ViewTypeButtonGroup from '@/components/artist/ViewTypeButtonGroup';
 import PageTitle from '@/components/common/PageTitle';
 import { sortTypes, viewTypes } from '@/data/artists';
 import { useDebounce } from '@/hook/useDebounce';
-import { getAuthorList } from '@/lib/service/client/artists';
+import { getAuthorList } from '@/service/client/artists';
 import useArtistsStore from '@/store/artistsStore';
 import { darkMode, lightMode } from '@/styles/theme';
 import type { AuthorInfoWithName, SortCriteria } from '@/types';
