+++ service/client/gallery.ts
@@ -1,6 +1,7 @@
 import axios from 'axios';
 
 // 이세돌 공지사항 모든 글 가져오기
+// 아래처럼 lastpage를 받고 page를 주는 형태로 변경 요청
 export const getIsdNotice = async () => {
   const response = await axios.get<IsdNotice>(
     `${process.env.NEXT_PUBLIC_SERVER_URL}/isd_notice`
@@ -9,6 +10,7 @@ export const getIsdNotice = async () => {
 };
 
 // 키워드 갤러리 작품들 가져오기
+// {"lastPage": true, "total": 22011, "list": []}
 export const getKeywordGalleryArtworks = async ({
   query,
   sortType,
